[SSHD-812] Made SftpSubsystem more thread-safe
